fixed problems in jinja template

+++ data/crawler.py
@@ -43,7 +43,7 @@ class Crawler:
             parent_score_aggregation_dic = json.loads(parent_score_aggregation_tag.string)
             score_aggregation_dic = parent_score_aggregation_dic["aggregateRating"]
             avg_rating_score = float(score_aggregation_dic["ratingValue"].replace(',', '.'))
-            rating_count = float(score_aggregation_dic["ratingCount"].replace(',', '.'))
+            rating_count = int(score_aggregation_dic["ratingCount"].replace(',', '.'))
        
         except:
             return None
@@ -116,7 +116,7 @@ class Crawler:
         note_count_tag = big_note_tag.find(class_ = "user-note-count")
         regexp = re.compile(r"(.\d+) note.*")
         
-        return float(note_tag.string.replace(',', '.')), float(regexp.match(note_count_tag.string)[1].replace(',', '.'))
+        return float(note_tag.string.replace(',', '.')), int(regexp.match(note_count_tag.string)[1].replace(',', '.'))
 
 
     def __is_review_url_tag(self, tag):
